+++ MoreShipUpgrades/Managers/ContractManager.cs
@@ -15,7 +15,6 @@ using System.Linq;
 using System.Text;
 using Unity.Netcode;
 using UnityEngine;
-using static LethalLib.Modules.Items;
 
 namespace MoreShipUpgrades.Managers
 {